LPS-197423 Migrate widgetsReducer to typescript
