+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/actions/addFragmentEntryLinks.ts
@@ -44,6 +44,7 @@ export interface FragmentEntryLink<
 	icon: string;
 	masterLayout?: boolean;
 	name: string;
+	portletId?: string;
 	removed: boolean;
 	segmentsExperienceId: string;
 }
