+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/actions/updateWidgets.ts
@@ -16,7 +16,7 @@ export interface Widget {
 }
 
 export interface WidgetSet {
-	categories: string[];
+	categories: WidgetSet[] | null;
 	path: string;
 	portlets: Widget[];
 	title: string;
