+++ modules/apps/layout/layout-content-page-editor-web/types/src/main/resources/META-INF/resources/page_editor/app/actions/addFragmentEntryLinks.d.ts
@@ -41,6 +41,7 @@ export interface FragmentEntryLink<
 	icon: string;
 	masterLayout?: boolean;
 	name: string;
+	portletId?: string;
 	removed: boolean;
 	segmentsExperienceId: string;
 }
