+++ modules/apps/layout/layout-content-page-editor-web/types/src/main/resources/META-INF/resources/page_editor/app/actions/updateWidgets.d.ts
@@ -12,7 +12,7 @@ export interface Widget {
 	title: string;
 }
 export interface WidgetSet {
-	categories: string[];
+	categories: WidgetSet[] | null;
 	path: string;
 	portlets: Widget[];
 	title: string;
