+++ modules/apps/layout/layout-content-page-editor-web/types/src/main/resources/META-INF/resources/page_editor/app/reducers/mappingFieldsReducer.d.ts
@@ -1,0 +1,13 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import addMappingFields, {
+	MappingFieldFieldSet,
+} from '../actions/addMappingFields';
+export declare const INITIAL_STATE: Record<string, MappingFieldFieldSet[]>;
+export default function mappedInfoItemsReducer(
+	mappingFields: Record<string, MappingFieldFieldSet[]> | undefined,
+	action: ReturnType<typeof addMappingFields>
+): Record<string, MappingFieldFieldSet[]>;
