+++ modules/apps/layout/layout-content-page-editor-web/types/src/main/resources/META-INF/resources/page_editor/app/reducers/widgetsReducer.d.ts
@@ -1,0 +1,13 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import toggleWidgetHighlighted from '../actions/toggleWidgetHighlighted';
+import updateWidgets, {WidgetSet} from '../actions/updateWidgets';
+export default function widgetsReducer(
+	widgets: WidgetSet[] | undefined,
+	action:
+		| ReturnType<typeof toggleWidgetHighlighted>
+		| ReturnType<typeof updateWidgets>
+): WidgetSet[] | null;
